Separeted content from the FriendsPage to a FriendsContainer Component.
